Starting CI
Added the jira-sync workflow, and a test workflow to check formatting,
run unit tests, and test a docker build (which also builds the code).

Cleared the old license from the boilerplate since it doesn't match
the current license, and we don't need a license header on every file.

Changed the Dockerfile to take a go version.