more robust status update
By the time updateStatus() is called, sometimes the CR in k8s has a
different resourceVersion so the status update fails. Usually nothing
has actually changed, so added a check do see if the Spec has actually
changed before proceeding in that case. Also made Renew required,
since its a bool and setting it to false meant it was being omitted
and showing up as a change. Also filtered out status update events, so
that updateStatus() doesn't cause extra Reconcile's.