Add HMAC based Secret data drift detection for VSS (#53)
Brings in a new feature to track the Secret.Data by storing the last
HMAC of the last Vault secret sync. Upon subsequent reconciliations the
"lastMAC" be checked against the current Secret's Data, if there is a
drift detected, the controller will sync the data from Vault to the
Secret.

We are also using this feature to prevent needless syncs. We can also
use it trigger rollout-restarts once the VSS supports that.

The integration tests for this feature are a bit zany, but they seem to
work well.

There are a few other unrelated/minor fixes included here.

* Add bit more context for test errors,

- bump the RefreshAfter to 7s to creation tests.
- use replace k8s.GetSecretE() with client.Client, the former was too
  chatty which made the test logs hard to read.
- misc fixes/updates

* HKDF tests, use simpler approach duplicate key creation.