Populate a pki-synced secret with tls fields
When a VaultPKISecret's Destination type is set to kubernetes.io/tls,
the `certificate` and `private_key` fields will also be copied to
`tls.crt` and `tls.key` fields in the k8s Secret.

Also added a sample with instructions in the README, and some kind
configs to expose an ingress port to the host.