Add support for installing the copywrite tool
The build workflow only runs on main, and executes the manifests make
target which will in turn wipe out the license headers of any generated
yaml files. This action breaks the check 'git porcelain' check, and thus
fails the build.

That is just one use-case. The other is for OSS contributions, we want
to ensure that those do not introduce the issue mentioned above. So, we
can compromise by adding a new make target: copywrite, which ensures
that everyone is using the same version.