Add dependabot and update GHA (#69)
Adds a dependabot config to check weekly for version updates for
go.mod and docker images, and check weekly for just security updates
to github actions. Also changed the usage of setup-go in the workflows
to use go-version-file to read the go version directly from
.go-version, instead of a separate step to read the file. The yaml
update in helm-reference-gen was just because it was flagged by trivy
with a CVE.

Also updates go to 1.20.2.

Also pins GHAs to trusted sha's, and changed the bats install to use
npm, since mig4/setup-bats isn't trusted yet, and hasn't been updated
in a while.