Add support for cloning a namespaced Vault Client (#157)
Client's can now be cloned for use in namespaces other than the one
configured in the VaultAuth. A new sub-cache has been added to the
clientCache. The sub-cache is responsible for caching cloned Clients.
A cached clone is pruned from the cache whenever its parent is evicted
or removed from the main cache. The spec of the ClientCacheKey has been
extended to support a new "clone" suffix which differentiates it from a
"parent" client. Client clones share the parent's token.

* Drop unused Namespace from ClientOptions