dynamic secrets: RenewalPercent and truncated renewals (#170)
Adds a new RenewalPercent field to VaultDynamicSecret for specifying
at what percentage of a lease the renewal should happen, defaults to
67 with an allowed range of 0-90.

Adds a new horizon function that computes the horizon as the specified
percentage of the lease, plus jitter of up to 10% of the lease, so
that the horizon falls after the specified percentage of the
lease. (This is inspired by the lifetimewatcher behavior from the
vault api.)

Also returns an error from lease renewal if the requested ttl was
capped, so that new credentials can be generated. Otherwise the
renewal ttl would be whatever is left on the max_ttl, which can become
very short before the renewal fails and and new credentials are
generated.
---------

Co-authored-by: Ben Ash <32777270+benashz@users.noreply.github.com>