Sync rendered chart with deployed resource
The chart specifies the skipTLSVerify value even if the value is
the default of false.  This causes the rendered yaml to be
different from the yaml reported when you read the resource from
kubernetes (via say kubectl get vaultconnection/default -oyaml).
This causes problems for orchestrators like Rancher as it sees
this as a modification.  This simple chagne leaves out the value
if it is the "default" - only including it if it is non-default
value of true, so the rendered manifest and reported yaml match.