Adding UBI image builds
Adds an image based on ubi-micro to Dockerfile, build directives in
Makefile, and image build jobs in the CI `build` job for dockerhub and
redhat (quay.io).

Also adds s390x and ppc64le to the arch list since they are supported
by both the distroless and ubi-micro base images.