Generate an OLM bundle with redhat-certified images (#310)
Adds a ClusterServiceVersion manifest and a default-openshift
kustomization for generating an OLM bundle that can be submitted to
Red Hat's certified-operators. The config/default-openshift kustomize
directory builds on config/default, but with the memory request and
limit increased for the manager, and an rbac-proxy image from the Red
Hat registry.