Handle invalid Client race after restoration (#400)
During the VSO restart the Vault Client is restored from the storage
cache, but it’s LifetimeWatcher has not yet renewed the client token,
which can lead to the next fetch from cache getting an invalid Client
(token expired) error. In this situation, the ClientFactory will evict
the client from the cache, which in turns calls client.Close(), which
has the negative side-effect of setting it’s wrapped api.Client to nil
thereby breaking any current Client references that are handling any
in-flight requests.