Add Go templating support
Extends VSO to support configuring Go templates that will be rendered to
the K8s Secret data. Template specs can be configured directly in a
syncable secret custom resource, or can be shared via references to
external ConfigMap resources. Templates can be composed of other
templates, and optionally rendered as a K8s Secret data value. The
rendered template name is used as the key to the rendered data.

Templating support also includes a large subset of the hermetic Sprig
text functions that can be used to handle various encoding/decoding
tasks, map manipulations, etc.

In addition to template support, it is now possible to filter top level
secret data fields based on include/exclude regex patterns. Filters can
be used in conjunction with templating, and are meant to augment the
overall templating experience. Template fields are not considered during
field filtering.