VDS: fix compute static-creds rotation horizon
Previously the jitter was a percentage on the static-creds rotation time
horizon. In the case where the rotation period/TTL was large e.g. 24h,
this would cause the rotation horizon extend beyond the rotation period.
This fix ensures that the added jitter is a 5% of the constant time
duration of 3s.