CCS: ensure invalid storage objects are deleted
Previously, it was possible for invalid client cache storage objects to
live forever. This could cause the ClientFactory to fail for valid
Client requests. This was most noticeable on a call to RestoreAll() when
VSO was first coming up.

Whenever an invalid cached client is found in storage it should be
discarded and replaced with a valid client instance.

Other fixes:
- Add more granular metrics for tracking storage deletions which are
  performed during storage pruning.
- Increase test coverage for cache_storage
- Remove unused EncryptionRequiredError