Rework the reference cache
Previously, the cache was keyed by referent object. This made things
simpler for referent management at the cost of extra memory usage and
some code complexity.

In order to mitigate the potential for encountering an out of memory
situation, the cache will map a referring object to its set of
references.

The previous implementation only supported adding or pruning referrers,
which would lead to orphaned references persisting in the cache. So,
Add() is replaced by Set(), which sets the set of references for each
referring object.