WIP: reconcile VDS instances on lifetimeWatcher errors
In the case where a lifetimeWatcher fails to renew the Vault client
lease, we want all related VDS instances to be synced. This helps to
mitigate the issue where an external revocation of the client token
causes the issued secret leases to be revoked. In that case VSO would
have no idea that the token has been revoked.

The ideal TTL for the client token should be relatively short, e.g. 1m,
so as to trigger the lifetimeWatcher earlier.

In the future, VSO will be able to subscribe to lease revocation events
from Vault. In that case, VSO will be able to perform the sync
immediately.