Reconcile VDS instances on lifetimeWatcher done events (#665)
In the case where a lifetimeWatcher fails to renew the Vault client
lease, we want all related VDS instances to be synced. This helps to
mitigate the issue where an external revocation of the client token
causes the issued secret leases to be revoked. In that case VSO would
have no idea that the token has been revoked.

The ideal TTL for the client token should be relatively short, e.g. 1m,
so as to trigger the lifetimeWatcher earlier.

In the future, VSO will be able to subscribe to lease revocation events
from Vault. In that case, VSO will be able to perform the sync
immediately.

This fix adds a the ability for the ClientFactory call back into the VDS reconciler
whenever the lifetime watcher has completed. The VDS reconciler will then 
perform a force sync on all VDS instances having the Client's cache key.
The Vault client cache key and ID persisted to the VDS instances status.

Iterative enhancement:
* Replace the SyncController with source.Channel (#704)

Update the VaultDynamicSecretReconciler to rely on the
controller-runtime's source.Channel that is watched as a raw source.
This ensures that there is only one reconciler work queue required to
handle external reconciliation requests.

Other fixes:
* No longer restore all clients from storage (#684)

Previously, the client factory would restore all stored clients to warm
its client cache. Some issues with that:
- restore all is blocking and can lead to a failure of the VSO pod to
  become ready if there a large number of clients to be restored
- restore all is unnecessary, since the client factory will always
  attempt a restoration upon resource reconciliation. All resources are
  reconciled when VSO starts up.

* Test infra: install metrics-server (#694)

It is nice to have k8s cluster metrics during integration testing.