CachingClientFactory: support client taints (#717)
* CachingClientFactory: support client taints

The secret reconcilers have no way of advising the CachingClientFactory
that a Client might be invalid. Now a reconciler can taint the Client if
a Vault operation fails for some reason. The common case is where a
Vault request resulted in a 403 (forbidden) status code. In this case
the reconciler can taint the client so that the next call to factory for
the tainted client will have the factory validate that the client's
token is still valid by performing reaching out to Vault's lookup lookup
API. Client taints should be used sparingly, since they can increase the
number of requests to Vault.

* Test lease error responses.

* Core: call callbacks on invalidated Client (#769)

Previously, invalidated Clients were removed from the Client cache
without calling any of the registered ClientCallbacks. Now any callbacks
registered with ClientCallbackOnCacheRemoval be invoked when a client is
removed from cache.