Drop the tainted client gauge.
In hindsight this metric doesn't make sense at this point. It also
introduces another high cardinality metric that users probably don't
want to have to manage. This metric is probably better implemented as a
collector on the client cache itself, setting the total number of
tainted clients would probably be more meaningful.