Fix: encryption client deadlocking the factory (#868)
* Fix: encryption client deadlocking the factory

Setting up the storage encryption client deadlocks the client factory if
the client is deemed to be invalid. This causes all reconciliations to
block forever.

The fix is to move the client setup code out of
storageEncryptionClient(). This gets rid of the vestigial recursive call
to itself and allows for single locker.