VC: add spec.timeout for Vault requests (#862)
Extends the VaultConnection API to include the support for setting the
Vault request time. It will be applied to all Vault Clients for the
given VaultConnection.

* Enable extra debugging to the demo

* Fix: encryption client deadlocking the factory (#868)

Setting up the storage encryption client deadlocks the client factory if
the client is deemed to be invalid. This causes all reconciliations to
block forever.

The fix is to move the client setup code out of
storageEncryptionClient(). This gets rid of the vestigial recursive call
to itself and allows for single locker.