Helm vault namespace do not set if empty
When using server side diff with ArgoCD the default provided vault auths
for both transit and the default fail to apply.

This is because they are set to null when rendered by ArgoCD because the value has not been quoted to
take into account for empty strings.

ArgoCD will render the namespace as

```yaml
namespace: null
```

which fails to apply because the CRD says it must be a string.

This fix drops the namespace if its not set, which will fix this error
for people using argocd with server side diff